+++ src/main/java/com/kuna_backend/dtos/cart/CartItemDto.java
@@ -5,8 +5,8 @@ import com.kuna_backend.models.Product;
 import com.kuna_backend.models.ProductVariation;
 
 public class CartItemDto {
-    private Integer id;
-    private Integer quantity;
+    private long id;
+    private int quantity;
     private Product product;
     private ProductVariation productVariation;
 
@@ -33,19 +33,19 @@ public class CartItemDto {
                 '}';
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public Integer getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
