+++ src/main/java/com/kuna_backend/repositories/CartRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CartRepository extends JpaRepository<Cart, Integer> {
+public interface CartRepository extends JpaRepository<Cart, Long> {
     List<Cart> findAllByClientOrderByCreatedAtDesc(Client client);
 
     List<Cart> deleteByClient(Client client);
